requirements.txt
python-telegram-bot[all]==22.5
APScheduler==3.11.2
fastapi==0.115.*
uvicorn[standard]==0.34.*
jinja2==3.1.*
httpx>=0.27
